Update README.md
Fixed some minor grammatical mistakes. Please DELETE this pull request as soon as you read it and make the corrections yourself so that there wouldn't be a record of me making the correction as I don't want to seem a grammar Natzi. I just want to help remove what might make people think lowly of you.